3.1.12 node jako klasa prywatna
