+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -37,5 +37,37 @@ public class IntLinkedList {
 		last = last.getPrev();
 		return ret;
 	}
+	
+	private class Node {
+
+		private final int value;
+		private Node prev;
+		private Node next;
+
+		public Node(final int i) {
+			value = i;
+		}
+
+		public int getValue() {
+			return value;
+		}
+
+		public Node getPrev() {
+			return prev;
+		}
+
+		public void setPrev(final Node prev) {
+			this.prev = prev;
+		}
+
+		public Node getNext() {
+			return next;
+		}
+
+		public void setNext(final Node next) {
+			this.next = next;
+		}
+
+	}
 
 }
